api/requirements.txt
fastapi
uvicorn
python-multipart
torch
diffusers
transformers
accelerate
opencv-python
Pillow
imageio[ffmpeg]